refactor added plugins
